Refactor tests to reduce duplication
