+++ test/unit/data/model/db/__init__.py
@@ -3,18 +3,7 @@ from collections import (
     namedtuple,
 )
 
-PRIVATE_OBJECT_STORE_ID = "my_private_data"
-
 MockTransaction = namedtuple("MockTransaction", "user")
-
-
-class MockObjectStore:
-
-    def is_private(self, object):
-        if object.object_store_id == PRIVATE_OBJECT_STORE_ID:
-            return True
-        else:
-            return False
 
 
 def verify_items(items, expected_items):
